+++ packages/flutter/lib/src/material/range_slider.dart
@@ -273,9 +273,8 @@ class RangeSlider extends StatefulWidget {
   /// There are two labels: one for the start thumb and one for the end thumb.
   ///
   /// Each label is rendered using the active [ThemeData]'s
-  /// [ThemeData.textTheme.body2] text style, with the
-  /// theme data's [ThemeData.colorScheme.onPrimaryColor]. The label's text
-  /// style can be overridden with [SliderThemeData.valueIndicatorTextStyle].
+  /// [ThemeData.accentTextTheme.body2] text style, and can be overridden
+  /// by changing the [SliderThemeData.valueIndicatorTextStyle].
   ///
   /// If null, then the value indicator will not be displayed.
   ///
@@ -993,7 +992,7 @@ class _RenderRangeSlider extends RenderBox with RelayoutWhenSystemFontsChangeMix
   }
 
   double _discretize(double value) {
-    double result = value.clamp(0.0, 1.0) as double;
+    double result = value.clamp(0.0, 1.0);
     if (isDiscrete) {
       result = (result * divisions).round() / divisions;
     }
@@ -1005,7 +1004,7 @@ class _RenderRangeSlider extends RenderBox with RelayoutWhenSystemFontsChangeMix
   }
 
   void _startInteraction(Offset globalPosition) {
-    final double tapValue = _getValueFromGlobalPosition(globalPosition).clamp(0.0, 1.0) as double;
+    final double tapValue = _getValueFromGlobalPosition(globalPosition).clamp(0.0, 1.0);
     _lastThumbSelection = sliderTheme.thumbSelector(textDirection, values, tapValue, _thumbSize, size, 0);
 
     if (_lastThumbSelection != null) {
@@ -1408,10 +1407,10 @@ class _RenderRangeSlider extends RenderBox with RelayoutWhenSystemFontsChangeMix
   }
 
   double _increaseValue(double value) {
-    return (value + _semanticActionUnit).clamp(0.0, 1.0) as double;
+    return (value + _semanticActionUnit).clamp(0.0, 1.0);
   }
 
   double _decreaseValue(double value) {
-    return (value - _semanticActionUnit).clamp(0.0, 1.0) as double;
+    return (value - _semanticActionUnit).clamp(0.0, 1.0);
   }
 }
